+++ errands/state.py
@@ -12,6 +12,7 @@ if TYPE_CHECKING:
     from errands.lib.notifications import ErrandsNotificationsDaemon
     from errands.widgets.loading_page import ErrandsLoadingPage
     from errands.widgets.shared.task_toolbar import (
+        DetailWindow,
         ErrandsDateTimeWindow,
         ErrandsAttachmentsWindow,
         ErrandsNotesWindow,
@@ -72,12 +73,14 @@ class State:
     # Attachments window
     attachments_window: ErrandsAttachmentsWindow | None = None
 
+    detail_window: DetailWindow | None = None
     priority_window: PriorityWindow | None = None
     tag_window: TagWindow | None = None
 
     @classmethod
     def init(cls) -> None:
         from errands.widgets.shared.task_toolbar import (
+            DetailWindow,
             ErrandsNotesWindow,
             ErrandsDateTimeWindow,
             ErrandsAttachmentsWindow,
@@ -85,6 +88,7 @@ class State:
             TagWindow,
         )
 
+        cls.detail_window = DetailWindow()
         cls.notes_window = ErrandsNotesWindow()
         cls.datetime_window = ErrandsDateTimeWindow()
         cls.attachments_window = ErrandsAttachmentsWindow()
