+++ errands/widgets/shared/task_toolbar/__init__.py
@@ -3,6 +3,7 @@ from errands.widgets.shared.task_toolbar.attachments_window import (
     ErrandsAttachmentsWindow,
 )
 from errands.widgets.shared.task_toolbar.datetime_window import ErrandsDateTimeWindow
+from errands.widgets.shared.task_toolbar.detail_window import DetailWindow
 from errands.widgets.shared.task_toolbar.notes_window import ErrandsNotesWindow
 from errands.widgets.shared.task_toolbar.priority_window import PriorityWindow
 from errands.widgets.shared.task_toolbar.tag_window import TagWindow
