+++ errands/widgets/shared/task_toolbar/toolbar.py
@@ -6,15 +6,12 @@ from __future__ import annotations
 from datetime import datetime
 from typing import TYPE_CHECKING
 
-from gi.repository import Adw, Gio, GLib, GObject, Gtk  # type:ignore
+from gi.repository import Adw, Gio, GLib, Gtk  # type:ignore
 
-from errands.lib.logging import Log
-from errands.lib.sync.sync import Sync
 from errands.lib.utils import get_human_datetime
 from errands.state import State
-from errands.widgets.shared.color_selector import ErrandsColorSelector
 from errands.widgets.shared.components.boxes import ErrandsBox
-from errands.widgets.shared.components.buttons import ErrandsButton, ErrandsCheckButton
+from errands.widgets.shared.components.buttons import ErrandsButton
 
 
 if TYPE_CHECKING:
@@ -84,61 +81,12 @@ class ErrandsTaskToolbar(Gtk.FlowBox):
             on_click=lambda *_: State.attachments_window.show(self.task),
         )
 
-        # Menu button
-        menu_top_section: Gio.Menu = Gio.Menu()
-        menu_colors_item: Gio.MenuItem = Gio.MenuItem()
-        menu_colors_item.set_attribute_value("custom", GLib.Variant("s", "color"))
-        menu_top_section.append_item(menu_colors_item)
-
-        menu: Gio.Menu = Gio.Menu()
-        menu.append_section(None, menu_top_section)
-
-        menu_bottom_section: Gio.Menu = Gio.Menu()
-        menu_created_item: Gio.MenuItem = Gio.MenuItem()
-        menu_created_item.set_attribute_value("custom", GLib.Variant("s", "created"))
-        menu_bottom_section.append_item(menu_created_item)
-        menu_changed_item: Gio.MenuItem = Gio.MenuItem()
-        menu_changed_item.set_attribute_value("custom", GLib.Variant("s", "changed"))
-        menu_bottom_section.append_item(menu_changed_item)
-        menu.append_section(None, menu_bottom_section)
-
-        popover_menu = Gtk.PopoverMenu(menu_model=menu)
-
-        # Colors
-        self.color_selector: ErrandsColorSelector = ErrandsColorSelector(
-            on_color_selected=self.__on_accent_color_selected
-        )
-        popover_menu.add_child(self.color_selector, "color")
-
-        # Created label
-        self.created_label = Gtk.Label(
-            label=_("Created:"),
-            halign=Gtk.Align.START,
-            margin_bottom=6,
-            margin_end=6,
-            margin_start=6,
-            css_classes=["caption-heading"],
-        )
-        popover_menu.add_child(self.created_label, "created")
-
-        # Changed label
-        self.changed_label = Gtk.Label(
-            label=_("Changed:"),
-            halign=Gtk.Align.START,
-            margin_end=6,
-            margin_start=6,
-            css_classes=["caption-heading"],
-        )
-        popover_menu.add_child(self.changed_label, "changed")
-
-        menu_btn: Gtk.MenuButton = Gtk.MenuButton(
-            popover=popover_menu,
+        detail_btn: ErrandsButton = ErrandsButton(
+            tooltip_text=_("Details"),
             icon_name="errands-more-symbolic",
             css_classes=["flat"],
-            valign=Gtk.Align.CENTER,
-            tooltip_text=_("More"),
+            on_click=lambda *_: State.detail_window.show(self.task),
         )
-        menu_btn.connect("notify::active", self._on_menu_toggled)
 
         self.append(
             ErrandsBox(
@@ -149,7 +97,7 @@ class ErrandsTaskToolbar(Gtk.FlowBox):
                     self.priority_btn,
                     tag_btn,
                     self.attachments_btn,
-                    menu_btn,
+                    detail_btn,
                 ],
             )
         )
@@ -190,35 +138,3 @@ class ErrandsTaskToolbar(Gtk.FlowBox):
         self.attachments_btn.remove_css_class("accent")
         if len(self.task.task_data.attachments) > 0:
             self.attachments_btn.add_css_class("accent")
-
-    def __on_accent_color_selected(
-        self, _color_selector, btn: ErrandsCheckButton, color: str
-    ) -> None:
-        if not btn.get_active() or self.task.block_signals:
-            return
-        Log.debug(f"Task: change color to '{color}'")
-        if color != self.task.task_data.color:
-            self.task.update_props(
-                ["color", "synced"], [color if color != "none" else "", False]
-            )
-            self.task.update_color()
-            Sync.sync()
-
-    def _on_menu_toggled(self, btn: Gtk.MenuButton, active: bool) -> None:
-        if not btn.get_active():
-            return
-
-        # Update dates
-        created_date: str = datetime.fromisoformat(
-            self.task.task_data.created_at
-        ).strftime("%Y.%m.%d %H:%M:%S")
-        changed_date: str = datetime.fromisoformat(
-            self.task.task_data.changed_at
-        ).strftime("%Y.%m.%d %H:%M:%S")
-        self.created_label.set_label(_("Created:") + " " + created_date)
-        self.changed_label.set_label(_("Changed:") + " " + changed_date)
-
-        # Update color
-        self.task.block_signals = True
-        self.color_selector.select_color(self.task.task_data.color)
-        self.task.block_signals = False
